pyproject.toml
[project]
name = "aviation-incident-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.8",
    "hdbscan>=0.8.42",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.2",
    "langchain-mistralai>=1.1.2",
    "langchain-text-splitters>=1.1.2",
    "loguru>=0.7.3",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pydantic>=2.13.2",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "streamlit>=1.56.0",
    "umap-learn>=0.5.12",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "litecli>=1.17.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
